reduced code duplication in GE pages by making a new component
