+++ pages/serviceHealthIssue.js
@@ -1,71 +1,8 @@
-import React, { Component } from "react";
-import PropTypes from "prop-types";
-import { connect } from "react-redux";
-import withI18N from "../lib/i18nHOC";
-import Layout from "../components/layout";
-import GuidedExperience from "../components/guided_experience";
-import GuidedExperienceProfile from "../components/guided_experience_profile";
+import React from "react";
+import GuidedExperiencePage from "../components/guided_experience_page";
 
-const section = "serviceHealthIssue";
-
-export class HealthIssue extends Component {
-  getTooltip = question => {
-    if (this.props.t("current-language-code") === "en") {
-      return question["tooltip_english"];
-    } else {
-      return question["tooltip_french"];
-    }
-  };
-
-  render() {
-    const { t, i18n, store, reduxState, url } = this.props;
-    const question = reduxState.questions.filter(
-      x => x.variable_name === section
-    )[0];
-    const pageTitle =
-      t("current-language-code") === "en"
-        ? question.guided_experience_page_title_english
-        : question.guided_experience_page_title_french;
-
-    return (
-      <Layout
-        i18n={i18n}
-        t={t}
-        hideNoscript={false}
-        title={pageTitle}
-        skipLink="#mainContent"
-      >
-        <GuidedExperience
-          id={section}
-          helperText={this.getTooltip(question)}
-          t={t}
-          url={url}
-          store={store}
-        >
-          <GuidedExperienceProfile
-            t={t}
-            selectorType={section}
-            store={store}
-            options={question["multiple_choice_options"]}
-          />
-        </GuidedExperience>
-      </Layout>
-    );
-  }
-}
-
-const mapStateToProps = reduxState => {
-  return {
-    reduxState: reduxState
-  };
+let ServiceHealthIssue = props => {
+  return <GuidedExperiencePage section="serviceHealthIssue" {...props} />;
 };
 
-HealthIssue.propTypes = {
-  reduxState: PropTypes.object,
-  i18n: PropTypes.object.isRequired,
-  t: PropTypes.func.isRequired,
-  url: PropTypes.object.isRequired,
-  store: PropTypes.object
-};
-
-export default connect(mapStateToProps)(withI18N(HealthIssue));
+export default ServiceHealthIssue;
